requirements.txt
flask
flask-login
flask-ldap3-login
requests
reportlab
python-dotenv
openstacksdk
